some warnings suppressed

+++ src/activations.cpp
@@ -10,12 +10,12 @@ vector<long double> softmax(vector<long double> input) {
     long double exp_sum = 0;
 
     // sum exponentials
-    for (int i=0; i<input.size(); i++) {
+    for (int i=0; i<(int)input.size(); i++) {
         exp_sum += exp(input[i]);
     }
 
     // calculating output softmax array
-    for (int i=0; i<input.size(); i++) {
+    for (int i=0; i<(int)input.size(); i++) {
         output.push_back (exp (input[i]) / exp_sum);
     }
